Move validateAlias to utils (#1195)
It's a util function pulled from a few different components and pages.
